+++ src/CssHexColor/_fixtures/CssHexColorDataDefinitionFixtures.ts
@@ -34,7 +34,7 @@
 
 import { HashMap } from "@safelytyped/core-types";
 import { CSS_EXTENDED_COLORS_TO_HEX } from "@safelytyped/css-color";
-import { DARK_COLORS, DULL_COLORS, LIGHT_COLORS } from "../../inspectors/_fixtures/colorShades";
+import { DARK_COLORS, MIDTONE_COLORS, LIGHT_COLORS } from "../../inspectors/_fixtures/colorShades";
 
 export const ValidCssHexColorDefinitions = [
     // we use Set() to dedupe the color definitions
@@ -42,7 +42,7 @@ export const ValidCssHexColorDefinitions = [
         ...HashMap.values(CSS_EXTENDED_COLORS_TO_HEX),
         ...LIGHT_COLORS,
         ...DARK_COLORS,
-        ...DULL_COLORS
+        ...MIDTONE_COLORS
 
         // add additional values here
     ]),