+++ src/inspectors/tonality.ts
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2024-present Ganbaro Digital Ltd
+// Copyright (c) 2025-present Ganbaro Digital Ltd
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -32,17 +32,31 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export const LIGHT_COLORS = [
-    "#fff",
-    "#ffffff",
-];
+import type { AnyCssColor } from "../CssColor/AnyCssColor.type";
+import { luma } from "./luma";
 
-export const DARK_COLORS = [
-    "#000",
-    "#000000",
-];
+/**
+ * Tonality is a set of valid return values from the {@link tonality} function
+ */
+export type Tonality = "light" | "dark" | "midtone";
 
-export const MIDTONE_COLORS = [
-    "#888",
-    "#888888",
-];
+/**
+ * tonality() uses the {@link luma} to work out whether the given `input`
+ * color is "light", "dark" or "midtone"
+ *
+ * @param input - the color to examine
+ * @returns the calculated tonality
+ */
+export function tonality(input: AnyCssColor): Tonality
+{
+    const y = luma(input);
+
+    if (y < 100) {
+        return "dark";
+    }
+    if (y > 155) {
+        return "light";
+    }
+
+    return "midtone";
+}